refactor: extracting simpleServer as it is useful for process test
